refactor(input): 패키지 이동
